Update AZ-104.md
Added my OpenAI AZ-104 Exam Proctor.  This is a free exam testing taking environment available if you have an OpenAI Plus account.  Not only does it test you, you can ask for details to explain topics and give deeper reasons as to why an answer is either right or wrong.